Implémentation des méthodes encrypter/décrypter.

+++ src/commun/LigneFacture.java
@@ -4,6 +4,8 @@ import java.io.Serializable;
 import java.text.NumberFormat;
 import java.util.Locale;
 
+import encryption.AES;
+
 /**
  * Classe représentant une ligne de la facture dans le TableView du POS
  * 
@@ -149,23 +151,23 @@ public class LigneFacture implements Serializable, Cryptable {
 	public void encrypter(String cle) {
 		this.prixUnitaire = AES.toFloat(AES.encrypter(cle, AES.fromFloat(this.prixUnitaire)));
 		this.prix = AES.toFloat(AES.encrypter(cle, AES.fromFloat(this.prix)));
-		this.quantite = AES.toFloat(AES.encrypter(cle, AES.fromInt(this.quantite)));
+		this.quantite = AES.toInt(AES.encrypter(cle, AES.fromInt(this.quantite)));
 		this.nom = AES.encrypter(cle, this.nom);
 		this.prixString = AES.encrypter(cle, this.prixString);
-		
+
 		this.produit.encrypter(cle);
-		
+
 	}
 
 	@Override
 	public void decrypter(String cle) {
 		this.prixUnitaire = AES.toFloat(AES.decrypter(cle, AES.fromFloat(this.prixUnitaire)));
 		this.prix = AES.toFloat(AES.decrypter(cle, AES.fromFloat(this.prix)));
-		this.quantite = AES.toFloat(AES.decrypter(cle, AES.fromInt(this.quantite)));
+		this.quantite = AES.toInt(AES.decrypter(cle, AES.fromInt(this.quantite)));
 		this.nom = AES.decrypter(cle, this.nom);
 		this.prixString = AES.decrypter(cle, this.prixString);
-		
+
 		this.produit.decrypter(cle);
-		
+
 	}
 }
